Create container and mini containers for items
Originally did not include text-align: center in body so made the items
flex instead which was unnecessary. Simply make the original container
flex for the items to be displayed side by side and make formatting
changes and center them instead of space-around such that all of them
are clumped at the center of the page.