Apply flex-shrink: 0 to prevent circle collapse
Initially made new container a flex object as well to facilitate the
formation of the right hand side elements but it was not needed as the
are fine being block display types and buttons by itself are inline so
making it flex would only lead it to fill the available space.

Understand that if you can use less flex box, use less as it keeps the
code concise and succinct.