Use display: flex on various divs
Still unclear on where I can use display: flex and flex: 1 and so on,
need to reread and reestablish where and which elements would benefit
from using the properties and in essence how it works, like in this
making body display: flex and flex-direction: column sets the whole
page into a flexbox. Then, display: flex in the header helps to center
it despite no other elements benefitting it, so you can remember that
display: flex can be used in such cases as well.

flex-wrap should be used on the container that you want the wrapping to
happen and not on the item that should be wrapped. flex: 1 will stretch
the element along the cross-axis.